+++ Net_Core/DataAccess/SQLServer/DAccount.cs
@@ -70,13 +70,13 @@ public class DAccounts : AbstractDBInteraction, IAccount
         return await base.GetDataRowAsync(mStoredProcedure, mParameters);
     }
 
-    async Task<DataRow> IAccount.GetAccountByVerificationToken()
+    async Task<DataSet> IAccount.GetAccountByVerificationToken()
     {
         String mStoredProcedure = "[ZGWSecurity].[Get_Account_By_Verification_Token]";
         SqlParameter[] mParameters = [
             GetSqlParameter("@P_VerificationToken", this.Cleanup(m_Profile.VerificationToken), ParameterDirection.Input)
         ];
-        return await base.GetDataRowAsync(mStoredProcedure, mParameters);
+        return await base.GetDataSetAsync(mStoredProcedure, mParameters);
     }
 
     async Task<DataTable> IAccount.GetAccounts()